refactor: ディレクトリ構造の変更2
