+++ src/pages/bookshelf/BookshelfGrid/BookCard.tsx
